pyproject.toml
[tool.poetry]
name = "terraglamp"
version = "0.1.0"
description = ""
authors = ["akorzunin <akorzunin123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
motor = "^3.1.2"
pydantic = {extras = ["email"], version = "^1.10.7"}
phonenumbers = "^8.13.11"
starlette-admin = "^0.8.2"
tinydb = "^4.7.1"
fastapi-mail = "^1.2.8"
aiogram = {url = "https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"
motor-types = "^1.0.0b1"
ruff = "^0.0.272"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = 3.11
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true


[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
